Add Lazy::into_value
This API allows to move a T out of `Lazy<T>`.
Note that it requires an owned access to a `Lazy` -- there's no way to
reset a lazy to uninit state via `&mut Lazy<T>`, as the init function
is gone. In other words, `fn take(this: &mut Lazy<T>)` is an
impossible object.